+++ stocky-client/src/components/login/Main.css
@@ -9,6 +9,10 @@ a:hover {
   color: #fff;
 }
 
+.btn-link {
+  color: #fff;
+}
+
 /* Custom default button */
 .btn-secondary,
 .btn-secondary:hover,
